Creating 3D Engine - Rendering 3D Model

+++ Tutorial/Texture.cpp
@@ -16,7 +16,7 @@ Texture::Texture(const wchar_t* full_path)
 		D3D11_SHADER_RESOURCE_VIEW_DESC desc = {};
 		desc.Format = image_data.GetMetadata().format;
 		desc.ViewDimension = D3D11_SRV_DIMENSION_TEXTURE2D;
-		desc.Texture2D.MipLevels = image_data.GetMetadata().mipLevels;
+		desc.Texture2D.MipLevels = static_cast<UINT>(image_data.GetMetadata().mipLevels);
 		desc.Texture2D.MostDetailedMip = 0;
 
 		Engine::GetInstance()->GetGraphics()->GetD3DDevice()->CreateShaderResourceView(m_texture, &desc, &m_shader_resource_view);
